backend/requirements.txt
fastapi>=0.94.0
uvicorn>=0.21.0
httpx>=0.24.1
requests>=2.28.2
redis>=4.6.0
kombu>=5.3.1
python-dotenv>=1.0.0
python-multipart>=0.0.6
pydantic>=1.10.6
pymongo>=4.4.0
motor>=3.2.0

numpy>=1.26.0
pyarrow>=14.0.0
pandas>=2.1.0

cryptography>=41.0.1
passlib>=1.7.4
PyJWT>=2.7.0

# Optional: include this only if your integrations require AWS auth for pymongo
# pymongo-auth-aws>=1.1.0

# Include these if you use them in your integrations or logging
# loguru>=0.6.0
# python-json-logger>=2.0.7

# Add any other packages you use directly, otherwise keep this minimal


# aiohttp==3.8.4
# aiosignal==1.3.1
# amqp==5.1.1
# anyio==3.6.2
# appnope==0.1.3
# apscheduler==3.10.1
# argon2-cffi==21.3.0
# argon2-cffi-bindings==21.2.0
# arrow==1.2.3
# asteval==0.9.31
# asttokens==2.2.1
# async-timeout==4.0.2
# attrs==22.2.0
# backcall==0.2.0
# bcrypt==4.0.1
# beautifulsoup4==4.11.2
# billiard==4.1.0
# bleach==6.0.0
# boto3==1.26.161
# botocore==1.29.161
# cachetools==5.3.1
# celery==5.3.1
# certifi==2023.5.7
# cffi==1.15.1
# charset-normalizer==3.1.0
# click==8.1.3
# click-didyoumean==0.3.0
# click-plugins==1.1.1
# click-repl==0.3.0
# comm==0.1.2
# contourpy==1.0.7
# cryptography==41.0.1
# cycler==0.11.0
# dataclasses-json==0.5.7
# debugpy==1.6.6
# decorator==5.1.1
# defusedxml==0.7.1
# dnspython==2.3.0
# ecdsa==0.18.0
# eventlet==0.33.3
# executing==1.2.0
# fastapi==0.94.0
# fastapi-pagination==0.12.5
# fastjsonschema==2.16.3
# fonttools==4.39.0
# fqdn==1.5.1
# frozenlist==1.3.3
# google-api-core==2.11.1
# google-api-python-client==2.96.0
# google-auth==2.22.0
# google-auth-httplib2==0.1.0
# google-auth-oauthlib==1.0.0
# googleapis-common-protos==1.60.0
# greenlet==2.0.2
# h11==0.14.0
# hiredis==2.2.3
# httpcore==0.17.3
# httplib2==0.22.0
# httptools==0.5.0
# httpx==0.24.1
# idna==3.4
# isoduration==20.11.0
# jedi==0.18.2
# Jinja2==3.1.2
# jmespath==1.0.1
# jsonpointer==2.3
# jsonschema==4.17.3
# kiwisolver==1.4.4
# kombu==5.3.1
# langchain==0.0.121
# loguru==0.6.0
# MarkupSafe==2.1.2
# marshmallow==3.19.0
# marshmallow-enum==1.5.1
# matplotlib==3.7.1
# matplotlib-inline==0.1.6
# mistune==2.0.5
# motor==3.2.0
# multidict==6.0.4
# mypy-extensions==1.0.0
# nbclassic==0.5.3
# nbclient==0.7.2
# nbconvert==7.2.9
# nbformat==5.7.3
# nest-asyncio==1.5.6
# networkx==3.1
# notebook==6.5.3
# notebook_shim==0.2.2
# numpy==1.26.0
# openai==0.27.2
# packaging==23.0
# pandas==1.5.3
# pandocfilters==1.5.0
# parso==0.8.3
# passlib==1.7.4
# pexpect==4.8.0
# pickleshare==0.7.5
# Pillow==9.4.0
# pinecone-client==2.2.1
# platformdirs==3.1.1
# prometheus-client==0.16.0
# prompt-toolkit==3.0.38
# protobuf==4.24.0
# psutil==5.9.4
# ptyprocess==0.7.0
# pure-eval==0.2.2
# pyarrow==14.0.0
# pyasn1==0.5.0
# pyasn1-modules==0.3.0
# pycparser==2.21
# pycurl==7.45.2
# pydantic==1.10.6
# Pygments==2.14.0
# PyJWT==2.7.0
# pymongo==4.4.0
# pymongo-auth-aws==1.1.0
# pymongoarrow==0.7.0
# pymongocrypt==1.6.1
# pyparsing==3.0.9
# pyrsistent==0.19.3
# python-dateutil==2.8.2
# python-dotenv==1.0.0
# python-jose==3.3.0
# python-json-logger==2.0.7
# python-multipart==0.0.6
# pytz==2022.7.1
# PyYAML==6.0
# pyzmq==25.0.0
# qtconsole==5.4.0
# QtPy==2.3.0
# redis==4.6.0
# requests==2.28.2
# rfc3339-validator==0.1.4
# rfc3986-validator==0.1.1
# rsa==4.9
# s3transfer==0.6.0
# Send2Trash==1.8.0
# six==1.16.0
# sniffio==1.3.0
# soupsieve==2.4
# SQLAlchemy==1.4.47
# stack-data==0.6.2
# starlette==0.26.0.post1
# stripe==5.5.0
# tenacity==8.2.2
# terminado==0.17.1
# tinycss2==1.2.1
# tornado==6.2
# tqdm==4.65.0
# traitlets==5.9.0
# typing-inspect==0.8.0
# typing_extensions==4.5.0
# tzdata==2023.3
# tzlocal==5.0.1
# uri-template==1.2.0
# uritemplate==4.1.1
# urllib3==1.26.16
# uvicorn==0.21.0
# uvloop==0.17.0
# vine==5.0.0
# watchfiles==0.19.0
# wcwidth==0.2.6
# webcolors==1.12
# webencodings==0.5.1
# websocket-client==1.5.1
# websockets==11.0.3
# widgetsnbextension==4.0.5
# yarl==1.8.2